Correction PK

+++ noisemodelling-emission/src/main/java/org/noise_planet/noisemodelling/emission/RailwayTrackParametersCnossos.java
@@ -51,7 +51,7 @@ public class RailwayTrackParametersCnossos {
         return isTunnel;
     }
 
-    public int getNTrack() {
+    public double getNTrack() {
         return nTrack;
     }
